+++ src/main/java/org/example/dao/CompanyDAO.java
@@ -4,12 +4,13 @@ import org.example.company.ITCompany;
 import org.example.company.employee.*;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface CompanyDAO {
-    Integer create(ITCompany itCompany);
-    ITCompany find(int id);
+    Long create(ITCompany itCompany);
+    ITCompany find(long id);
     void addDeveloper(Developer developer);
     void addPM(PM pm);
-    List<Employee<ITRole>> getEmployeesByRole(ITRole role, int company_id);
-    Employee<ITRole> findEmployee(int employee_id);
+    List<Employee<ITRole>> getEmployeesByRole(ITRole role, long company_id);
+    Employee<ITRole> findEmployee(long employee_id);
 }
